move application steps
